header file

+++ main.h
@@ -36,15 +36,21 @@ typedef struct buffer_s
 } buffer_t;
 
 /**
- * struct converter_s - A new type defining a converter struct.
+ * struct converter_s -A new type defining a converter struct.
  * @specifier: A character representing a conversion specifier.
  * @func: A pointer to a conversion function corresponding to specifier.
+ * va_list: vafiable argument lists.
+ * int: integer.
+ * char: character
+ * buffer_t: A type buffer.
+ * unsigned int: a positive number.
+ * unsigned char: character
  */
 typedef struct converter_s
 {
 	unsigned char specifier;
-	unsigned int (*func)(va_list, buffer_t *,
-			unsigned char, int, int, unsigned char);
+
+	unsigned int (*func)(va_list, buffer_t *, size_t char, int, int, size_t char);
 } converter_t;
 
 /**
@@ -93,6 +99,17 @@ unsigned char handle_flags(const char *flags, char *index);
 unsigned char handle_length(const char *modifier, char *index);
 int handle_width(va_list args, const char *modifier, char *index);
 int handle_precision(va_list args, const char *modifier, char *index);
+
+/**
+ * handle_specifiers - Matches a conversion specifier with
+ *                     a corresponding conversion function.
+ * specifier: A pointer to a potential conversion specifier.
+ * @int: integer
+ * @char: character
+ *
+ * Return: If a conversion function is matched - a pointer to the function.
+ *         Otherwise - NULL.
+ */
 unsigned int (*handle_specifiers(const char *specifier))(va_list, buffer_t *,
 		unsigned char, int, int, unsigned char);
